+++ app/lib/features/chat_ui_showcase/models/mocks/mock_user.dart
